initial conversion
mod loads but there are no items and no recipes... probably what I get
for trying to port someone else's code, especially code that I am not
entirely sure of as to what is going on in it. Anyhow ... I'm willing to
bet this is both why RebelKiethy hasn't responded to me at all because
he knows this is goofy as fuck code and that he surely doesn't want to
port it to 1.12.2 because of this.